requirements.txt
asyncio==3.4.3
async-lru==2.0.4
discord.py==2.3.1
Pillow==10.0.0
python-dotenv==1.0.0
twitchAPI==3.11.0
streamlink==6.0.0
opencv-python==4.8.0.74
pytesseract==0.3.10